+++ apps/framework/shared/application-version.ts
@@ -6,8 +6,8 @@ export type ApplicationVersion = {
 }
 
 export const applicationVersion: ApplicationVersion = {
-  referenceNumber: 186,
-  version: "1.0.186",
-  label: "v 1.0.186",
-  releaseTag: "v-1.0.186",
+  referenceNumber: 188,
+  version: "1.0.188",
+  label: "v 1.0.188",
+  releaseTag: "v-1.0.188",
 }